Refactored Inventory
Set up so that I don't need to maintain a complicated interrelationship of mats to blocks.  Now, each MyTile knows what it drops (string) and what it requires to build (string) and those strings are compared to all prefabs tagged "mat" for materials.  No more fiddly indexing.  I wasn't at all sure that it would work, but it does.